Remove form from content items in favour of pulling directly from settings
Removed setting the lever apply form ID on a lever post when it's
created via the background task in favour of pulling the form directly
from the settings. Benefit of this is the form can be changed in a
single place instead of having to update all the lever postings (which
isn't actually possible anyways).