Update fetching content item from referer
When an application is submitted, the action will use the URL referrer
to identify which job posting the user is applying to. However in a
recent change the URL for a job posting was changed to include the
ID of the job posting. To cater for this and any future changes, the
logic for calculating the posting URL now strips out the protocol
scheme, host and tenant url prefix and uses what's left over to
match the URL of a content item.